+++ src/pwshBedrock/Public/Reset-ModelTally.ps1
@@ -77,7 +77,10 @@ function Reset-ModelTally {
             'mistral.mistral-large-2402-v1:0',
             'mistral.mistral-large-2407-v1:0',
             'mistral.mixtral-8x7b-instruct-v0:1',
-            'stability.stable-diffusion-xl-v1'
+            'stability.stable-diffusion-xl-v1',
+            'stability.stable-image-ultra-v1:0',
+            'stability.stable-image-core-v1:0',
+            'stability.sd3-large-v1:0'
         )]
         [string]$ModelID,